requirements.txt
robotframework==4.1.3
robotframework-selenium2library==3.0.0
Flask==2.1.0
gunicorn==20.0.4
chromedriver-binary==102.0.5005.61
Jinja2==3.0.3
